[ALPS04655827] mm: fix possioble integer overflows
Coverty reports 3 possible integer overflows. Fix them
in this change.
Unintentional integer overflow:8246713, 8246676, 8246721

MTK-Commit-Id: a66469283a78cbf8f855ca0be392fae9f8145fec

CR-Id: ALPS04655827
Feature: [Module]Kernel Memory Management

+++ mm/page_alloc.c
@@ -6610,7 +6610,7 @@ void __init mem_init_print_info(const char *str)
 
 #ifdef CONFIG_MTK_MEMCFG
 		kernel_reserve_meminfo.available =
-			nr_free_pages() << PAGE_SHIFT;
+			(unsigned long long)nr_free_pages() << PAGE_SHIFT;
 		kernel_reserve_meminfo.total = physpages << PAGE_SHIFT;
 		kernel_reserve_meminfo.kernel_code = codesize;
 		kernel_reserve_meminfo.rwdata = datasize;
@@ -6618,9 +6618,11 @@ void __init mem_init_print_info(const char *str)
 		kernel_reserve_meminfo.init = init_data_size + init_code_size;
 		kernel_reserve_meminfo.bss = bss_size;
 		kernel_reserve_meminfo.reserved =
-			(physpages - totalram_pages) << PAGE_SHIFT;
+			(unsigned long long)(physpages - totalram_pages) <<
+			PAGE_SHIFT;
 #ifdef CONFIG_HIGHMEM
-		kernel_reserve_meminfo.highmem = totalhigh_pages << PAGE_SHIFT;
+		kernel_reserve_meminfo.highmem =
+			(unsigned long long)totalhigh_pages << PAGE_SHIFT;
 #endif
 #endif
 }
